pyproject.toml
[project]
name = "scraper-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-pagination>=0.15.10",
    "fastapi[standard]>=0.134.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "strawberry-graphql[fastapi]>=0.307.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "app"]


[dependency-groups]
dev = [
    "faker>=40.5.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
